listener: keep signal lane tests isolated
